Keep the Da Vinci narration contract suite free of prisma
CI caught what the local run could not: the contract-tests job has no
DATABASE_URL, and server/utils/prisma.ts throws at module load without one.
verifyDaVinciNarration.ts reached prisma transitively — through davinci.ts for
DAVINCI_DIMENSIONS, and through davinciNarration.ts's narrator lookup — so the
whole "Contract verifiers" job died on an import, not an assertion. Locally the
suite passed only because the shell had DATABASE_URL exported for vue-tsc.

Split the pure pieces out rather than granting the job a database:

  * server/utils/davinciDimensions.ts — the ten dimensions, DAVINCI_PASS_VALUE,
    and resolveOutcomeKey, with no imports at all. davinci.ts re-exports all
    four, so every existing importer is untouched.
  * loadRunNarrator moved from davinciNarration.ts into the narrate route,
    which already owns the request's data loading. That was the module's only
    prisma dependency; it now imports nothing but the dimension vocabulary.

Both contract suites verified with DATABASE_URL explicitly unset, matching the
CI environment. vue-tsc 0 errors, eslint and prettier clean.

The other two red checks on this PR (facet-catalog, facet-alias-smoke) are
unrelated to this branch — facet-catalog fails on
`plugins/20.facet-catalog.client.ts is missing required contract text:
hydrateAdventureBuilder`, and this diff touches no facet or plugin files.
